requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-social-oauth2==2.1.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pytest-django==4.5.2
python-dotenv==1.0.0
gunicorn==20.1.0